requirements.txt
numpy
pyyaml
gymnasium
matplotlib